Restructure panel and folder structure
Original structure had the panel displaying the contents of only the
active folder (like a traditional file explorer).

The new intention is for the panel to show all folders of the tab's
colour, then allow any folders to be toggled open/close showing their
contents (like a file tree).

This will allow drag+drop to move lines and folders with ease.